requirements.txt
pyodbc==4.0.39
python-pptx==0.6.21
python-dotenv==0.19.0 